Fixed bug in from source genaration Fixed bug in Decimal annotation cconverting long to double Fixed bug in method invokation

+++ src/main/java/jenerator/annotations/DecimalNumberGenerable.java
@@ -10,8 +10,8 @@ import java.lang.annotation.Target;
 @Target(FIELD)
 public @interface DecimalNumberGenerable {
 
-	long MAX_DEFAULT = Long.MAX_VALUE;
-	long MIN_DEFAULT = Long.MIN_VALUE;
+	double MAX_DEFAULT = Double.MAX_VALUE;
+	double MIN_DEFAULT = Double.MIN_VALUE;
 	int PRECISION = 2;
 
 	/**
